added text log functionality
Added a functionality that allows to see the results of the analysis
in a .txt file whose name is the same than the excel output file,
but adding "_logfile".

Also, added two conditionals that check there is at least 1 pdf file
to analyze and thet check if at least 1 successful analysis was made
over those pdf documents, in order to avoid errors in the program
due to empty dataframes.